refactor: RestaurantItem 컴포넌트 분리 및 Restaurant 파일 구조 재설정
